backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
openai
pydantic
python-dotenv
mcp
google-generativeai